+++ src/utils/transformDeep.ts
@@ -2,7 +2,17 @@ import { isPrimitive } from './utils';
 
 type ObjectTransformer = (target: object, source: object) => object;
 
-export function transformDeep(target: any, source: any, callback: ObjectTransformer, visited = new Set()): any {
+type PropertyPreTransformer = (target: object, source: object, key: string | symbol) => boolean;
+
+export class TransformOptions {
+
+    /**
+     * Return false to prevent further transformation of the property.
+     */
+    public propertyPreTransform: PropertyPreTransformer;
+}
+
+export function transformDeep(target: any, source: any, callback: ObjectTransformer, options = new TransformOptions(), visited = new Set()): any {
 
     // not traversing primitives
     if (isPrimitive(target) || isPrimitive(source))
@@ -16,10 +26,13 @@ export function transformDeep(target: any, source: any, callback: ObjectTransfor
     // transform children
     Object.keys(target).forEach(key => {
 
+        if (options.propertyPreTransform && (options.propertyPreTransform(target, source, key) === false))
+            return;
+
         // transform child
         var subTarget = target[key];
         var subSource = source[key];
-        var newSubTarget = transformDeep(subTarget, subSource, callback, visited);
+        var newSubTarget = transformDeep(subTarget, subSource, callback, options, visited);
 
         // assign only if changed
         if (newSubTarget !== subTarget) {